Separate Polymarket position transforms
