+++ src/features/polymarket/utils/transforms.ts
@@ -1,8 +1,5 @@
-import { getAddress } from 'viem';
-
-import { useCurrencyConversionStore } from '@/features/currency/stores/currencyConversionStore';
 import { getLeague } from '@/features/polymarket/leagues';
-import { type PolymarketPosition, type PolymarketTeamInfo, type RawPolymarketPosition } from '@/features/polymarket/types';
+import { type PolymarketTeamInfo } from '@/features/polymarket/types';
 import {
   type PolymarketEvent,
   type PolymarketMarket,
@@ -46,32 +43,6 @@ export async function processRawPolymarketEvent(event: RawPolymarketEvent, teams
     color,
     teams,
     league,
-  };
-}
-
-export async function processRawPolymarketPosition(
-  position: RawPolymarketPosition,
-  market: RawPolymarketMarket,
-  teams?: PolymarketTeamInfo[]
-): Promise<PolymarketPosition> {
-  const event = market.events[0];
-  const marketHasUniqueImage = market.icon !== event.icon;
-  const rawEventColor = await getImagePrimaryColor(event.icon);
-  const eventColor = { dark: getHighContrastColor(rawEventColor, true), light: getHighContrastColor(rawEventColor, false) };
-
-  return {
-    ...position,
-    proxyWallet: getAddress(position.proxyWallet),
-    clobTokenIds: market.clobTokenIds ? JSON.parse(market.clobTokenIds) : [],
-    outcomes: market.outcomes ? JSON.parse(market.outcomes) : [],
-    outcomePrices: market.outcomePrices ? JSON.parse(market.outcomePrices) : [],
-    nativeCurrency: {
-      currentValue: useCurrencyConversionStore.getState().convertToNativeCurrency(position.currentValue),
-      cashPnl: useCurrencyConversionStore.getState().convertToNativeCurrency(position.cashPnl),
-    },
-    market: processRawPolymarketMarket(market, eventColor),
-    marketHasUniqueImage,
-    teams,
   };
 }
